\name{RcppGO.Plot.Wrapper}
\alias{RcppGO.Plot.Wrapper}
\title{RcppGO.Plot.Wrapper()}
\usage{
  RcppGO.Plot.Wrapper(fn)
}
\arguments{
  \item{fn}{The objective function.}
}
\value{
  readable code for 'outer()'.
}
\description{This function modifies the user input 'fn'
  into the standard pattern required by the 'outer()'
  function.}
\author{
  Peter Kehler
}

